feat(algorithms): split dijkstra into files
boi got _heavy_
